ourochronos::ast

Enum OpCode

pub enum OpCode {

    Nop,
    Halt,
    Pop,
    Dup,
    Swap,
    Over,
    Rot,
    Depth,
    Pick,
    Add,
    Sub,
    Mul,
    Div,
    Mod,
    Neg,
    Not,
    And,
    Or,
    Xor,
    Shl,
    Shr,
    Eq,
    Neq,
    Lt,
    Gt,
    Lte,
    Gte,
    Oracle,
    Prophecy,
    PresentRead,
    Paradox,
    Pack,
    Unpack,
    Index,
    Store,
    Input,
    Output,

}

All operations available in OUROCHRONOS.

Variants

Nop

No operation.

Halt

Halt execution of the current epoch. Stack: ( โ )

Pop

Pop and discard the top of stack. Stack: ( a โ )

Dup

Duplicate the top of stack. Stack: ( a โ a a )

Swap

Swap the top two elements. Stack: ( a b โ b a )

Over

Copy the second element to the top. Stack: ( a b โ a b a )

Rot

Rotate the top three elements. Stack: ( a b c โ b c a )

Depth

Push the current stack depth. Stack: ( โ n )

Pick

Pick the nth element from deep in the stack and copy it to top. 0 PICK is equivalent to DUP. 1 PICK is equivalent to OVER. Stack: ( n โ v )

Add

Addition. Stack: ( a b โ a+b )

Sub

Subtraction. Stack: ( a b โ a-b )

Mul

Multiplication. Stack: ( a b โ a*b )

Div

Division (returns 0 if divisor is 0). Stack: ( a b โ a/b )

Mod

Modulo (returns 0 if divisor is 0). Stack: ( a b โ a%b )

Neg

Negation (twoโs complement). Stack: ( a โ -a )

Not

Bitwise NOT. Stack: ( a โ ~a )

And

Bitwise AND. Stack: ( a b โ a&b )

Or

Bitwise OR. Stack: ( a b โ a|b )

Xor

Bitwise XOR. Stack: ( a b โ a^b )

Shl

Left shift. Stack: ( a n โ a<<n )

Shr

Right shift (logical). Stack: ( a n โ a>>n )

Eq

Equal. Stack: ( a b โ a==b )

Neq

Not equal. Stack: ( a b โ a!=b )

Lt

Less than. Stack: ( a b โ a<b )

Gt

Greater than. Stack: ( a b โ a>b )

Lte

Less than or equal. Stack: ( a b โ a<=b )

Gte

Greater than or equal. Stack: ( a b โ a>=b )

Oracle

ORACLE: Read from anamnesis (the future). Pops an address, pushes the value from Anamnesis[address]. This is the mechanism for receiving information from the future. Stack: ( addr โ A[addr] )

Prophecy

PROPHECY: Write to present (fulfilling the future). Pops value and address, writes value to Present[address]. This is the mechanism for sending information to the past. Stack: ( value addr โ )

PresentRead

PRESENT_READ: Read from present memory (current epoch). Stack: ( addr โ P[addr] )

Paradox

PARADOX: Signal explicit inconsistency. Terminates the current epoch as paradoxical. In fixed-point search, this path is rejected. Stack: ( โ )

Pack

Pack n values into contiguous memory starting at base address. Stack: ( v1 v2 โฆ vn base n โ )

Unpack

Unpack n values from contiguous memory at base address. Stack: ( base n โ v1 v2 โฆ vn )

Index

Read from memory at computed index: base + index. Stack: ( base index โ P[base+index] )

Store

Store to memory at computed index: base + index. Stack: ( value base index โ )

Input

Read a value from input. Stack: ( โ value )

Output

Output a value. Stack: ( value โ )

Implementations

impl OpCode

pub fn name(&self) -> &'static str

Get the name of this opcode as it appears in source code.

pub fn stack_effect(&self) -> (usize, usize)

Get the stack effect: (inputs, outputs).